requirements.txt
streamlit
pdfplumber
pandas
openai
python-dotenv
plotly
anthropic
fpdf
reportlab
markdown
matplotlib
bs4
pillow